pyproject.toml
[project]
name = "gemini-orbiter"
version = "1.0.0"
description = "API Server built on Gemini Web"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "gemini-webapi>=1.12.1",
    "loguru>=0.7.0",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.34.1",
    "uvloop>=0.21.0",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"

[project.optional-dependencies]
dev = [
    "ruff>=0.11.7",
]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "W", "I", "RUF"]
lint.ignore = ["E501"] 

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
